share comment and export infrastructure
