+++ packages/react/package.json
@@ -86,7 +86,6 @@
     "emoji-picker-react": "^4.19.1",
     "fast-glob": "^3.3.2",
     "fflate": "^0.8.3",
-    "html-to-image": "^1.11.13",
     "lucide-react": "^1.25.0",
     "next-themes": "^0.4.6",
     "react": "^19.2.8",
