+++ packages/shared/src/client/index.ts
@@ -1,2 +1,9 @@
 export type { AssetEntry } from './assets.ts';
 export { deleteAsset, listAssets, renameAsset, uploadAsset } from './assets.ts';
+export {
+  exportFramesAsImagePptx,
+  type ImagePptxExportOptions,
+  type PptxExportProgress,
+  type PptxFrameCleanup,
+} from './export-pptx.ts';
+export { isFrameAnimationSettled, waitForDataWaitfor, waitForFonts } from './print-ready.ts';
