+++ packages/shared/src/editing/index.ts
@@ -1,2 +1,3 @@
 export { type AstNode, parseSource, walkAll, walkJsx } from './babel-walk.ts';
+export * from './comments.ts';
 export * from './slide-ops.ts';
